[reorg] Stage 2
- Merge `restate-network` into `restate-core`
- Merge NetworkError and NetworkSendError
