+++ crates/admin/src/service.rs
@@ -20,8 +20,8 @@ use tower::ServiceBuilder;
 use tracing::info;
 
 use restate_core::metadata_store::MetadataStoreClient;
+use restate_core::network::protobuf::node_svc::node_svc_client::NodeSvcClient;
 use restate_core::{cancellation_watcher, task_center, MetadataWriter};
-use restate_network::protobuf::node_svc::node_svc_client::NodeSvcClient;
 use restate_service_protocol::discovery::ServiceDiscovery;
 use restate_types::schema::subscriptions::SubscriptionValidator;
