+++ crates/core/src/network/mod.rs
@@ -8,10 +8,21 @@
 // the Business Source License, use of this software will be governed
 // by the Apache License, Version 2.0.
 
+mod connection;
+mod connection_manager;
 mod error;
+pub mod grpc_util;
+mod handshake;
 mod message_router;
+pub(crate) mod metric_definitions;
 mod network_sender;
+mod networking;
+pub mod protobuf;
+pub mod rpc_router;
 
+pub use connection::ConnectionSender;
+pub use connection_manager::ConnectionManager;
 pub use error::*;
 pub use message_router::*;
 pub use network_sender::*;
+pub use networking::Networking;
