+++ crates/core/src/network/protobuf.rs
