--- crates/network/Cargo.toml
@@ -1,46 +1,0 @@
-[package]
-name = "restate-network"
-version.workspace = true
-authors.workspace = true
-edition.workspace = true
-rust-version.workspace = true
-license.workspace = true
-publish = false
-
-[dependencies]
-restate-core = { workspace = true }
-restate-grpc-util = { workspace = true }
-restate-types = { workspace = true }
-
-anyhow = { workspace = true }
-async-trait = { workspace = true }
-bytes = { workspace = true }
-dashmap = { workspace = true }
-enum-map = { workspace = true }
-enumset = { workspace = true }
-futures = { workspace = true }
-http = { workspace = true }
-metrics = { workspace = true }
-once_cell = { workspace = true }
-pin-project = { workspace = true }
-prost = { workspace = true }
-prost-types = { workspace = true }
-rand = { workspace = true }
-static_assertions = { workspace = true }
-thiserror = { workspace = true }
-tokio = { workspace = true }
-tokio-stream = { workspace = true }
-tonic = { workspace = true, features = ["transport", "codegen", "prost", "gzip"] }
-tower = { workspace = true }
-tracing = { workspace = true }
-
-[build-dependencies]
-tonic-build = { workspace = true }
-
-[dev-dependencies]
-restate-core = { workspace = true, features = ["test-util"] }
-restate-test-util = { workspace = true }
-
-googletest = { workspace = true }
-tokio = { workspace = true, features = ["test-util"] }
-tracing-subscriber = { workspace = true }
