--- crates/network/src/lib.rs
@@ -1,22 +1,0 @@
-// Copyright (c) 2024 -  Restate Software, Inc., Restate GmbH.
-// All rights reserved.
-//
-// Use of this software is governed by the Business Source License
-// included in the LICENSE file.
-//
-// As of the Change Date specified in that file, in accordance with
-// the Business Source License, use of this software will be governed
-// by the Apache License, Version 2.0.
-
-mod connection;
-mod connection_manager;
-pub mod error;
-mod handshake;
-pub(crate) mod metric_definitions;
-mod networking;
-pub mod protobuf;
-pub mod rpc_router;
-
-pub use connection::ConnectionSender;
-pub use connection_manager::ConnectionManager;
-pub use networking::Networking;
