+++ crates/worker/src/partition/leadership/mod.rs
@@ -8,41 +8,42 @@
 // the Business Source License, use of this software will be governed
 // by the Apache License, Version 2.0.
 
-use crate::metric_definitions::PARTITION_HANDLE_LEADER_ACTIONS;
-use crate::partition::shuffle::{HintSender, Shuffle, ShuffleMetadata};
-use crate::partition::{shuffle, storage};
-use futures::future::OptionFuture;
-use futures::{future, StreamExt};
-use metrics::counter;
-use restate_core::network::NetworkSender;
-use restate_core::{
-    current_task_partition_id, metadata, task_center, ShutdownError, TaskId, TaskKind,
-};
-use restate_invoker_api::InvokeInputJournal;
-use restate_network::Networking;
-use restate_timer::TokioClock;
-use restate_types::net::ingress;
+mod action_collector;
+
 use std::fmt::Debug;
 use std::ops::RangeInclusive;
 use std::pin::Pin;
+
+use futures::future::OptionFuture;
+use futures::{future, StreamExt};
+use metrics::counter;
 use tokio::sync::mpsc;
 use tracing::{debug, trace, warn};
 
-mod action_collector;
-
-use crate::partition::action_effect_handler::ActionEffectHandler;
-use crate::partition::state_machine::Action;
-pub(crate) use action_collector::{ActionEffect, ActionEffectStream};
 use restate_bifrost::Bifrost;
+use restate_core::network::NetworkSender;
+use restate_core::network::Networking;
+use restate_core::{
+    current_task_partition_id, metadata, task_center, ShutdownError, TaskId, TaskKind,
+};
 use restate_errors::NotRunningError;
+use restate_invoker_api::InvokeInputJournal;
 use restate_partition_store::PartitionStore;
 use restate_storage_api::deduplication_table::EpochSequenceNumber;
+use restate_timer::TokioClock;
 use restate_types::identifiers::{InvocationId, PartitionKey};
 use restate_types::identifiers::{LeaderEpoch, PartitionId, PartitionLeaderEpoch};
+use restate_types::net::ingress;
 use restate_types::GenerationalNodeId;
 use restate_wal_protocol::timer::TimerKeyValue;
 
 use super::storage::invoker::InvokerStorageReader;
+use crate::metric_definitions::PARTITION_HANDLE_LEADER_ACTIONS;
+use crate::partition::action_effect_handler::ActionEffectHandler;
+use crate::partition::shuffle::{HintSender, Shuffle, ShuffleMetadata};
+use crate::partition::state_machine::Action;
+use crate::partition::{shuffle, storage};
+pub(crate) use action_collector::{ActionEffect, ActionEffectStream};
 
 type PartitionStorage = storage::PartitionStorage<PartitionStore>;
 type TimerService = restate_timer::TimerService<TimerKeyValue, TokioClock, PartitionStorage>;
