+++ tools/xtask/Cargo.toml
@@ -12,7 +12,6 @@ restate-admin = { workspace = true, features = ["options_schema"] }
 restate-bifrost = { workspace = true, features = ["options_schema", "test-util"] }
 restate-core = { workspace = true, features = ["test-util"] }
 restate-metadata-store = { workspace = true, features = ["test-util"] }
-restate-network = { workspace = true }
 restate-service-client = { workspace = true }
 restate-service-protocol = { workspace = true, features = ["discovery"]}
 restate-storage-query-datafusion = { workspace = true, features = ["table_docs"] }
